requirements.txt
Flask
pymongo
Flask-Caching
flasgger
Flask-Cors
python-dotenv
google-generativeai
gunicorn